requirements.txt
alembic==1.7.1
asgiref==3.4.1
click==8.0.1
ecdsa==0.17.0
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
importlib-resources==5.2.2
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.9.1
pyasn1==0.4.8
pydantic==1.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
zipp==3.5.0
